requirements.txt
alembic==1.16.5
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
dnspython==2.8.0
dominate==2.9.1
email-validator==2.3.0
Flask==3.1.2
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
gunicorn==23.0.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.2
python-dotenv==1.1.1
requests==2.32.5
SQLAlchemy==2.0.43
stripe==13.0.0
typing_extensions==4.15.0
urllib3==2.5.0
visitor==0.1.3
Werkzeug==3.1.3
WTForms==3.2.1